Milestone4/requirements.txt
streamlit
requests
python-dotenv
huggingface_hub
 PyJWT
 streamlit>=1.32.0
huggingface_hub>=0.21.0
groq>=0.9.0
huggingface_hub>=0.22.0
supabase>=2.0.0
requests>=2.31.0
sib-api-v3-sdk>=7.6.0
